backend/requirements.txt
# Web framework
fastapi==0.129.0
uvicorn[standard]==0.40.0
starlette==0.52.1

# Data validation
pydantic==2.12.5

# Environment
python-dotenv==1.2.1

# Vector store
chromadb==1.5.0

# Google Gemini (for text generation only)
google-genai==1.15.0

# Local embeddings (no API quota required)
sentence-transformers==5.2.2

# HTTP client
httpx==0.28.1

# Utilities
tqdm==4.67.3
requests==2.32.5
